add tests for different identity verification configs
