--- packages/tests-e2e/src/ui/chromium/signup/PasskeyAppendScreen-robustness.spec.ts
@@ -1,40 +1,0 @@
-import { test } from '../../../fixtures/UISignupTest';
-import { ScreenNames } from '../../../utils/constants';
-
-test.describe('PasskeyAppendScreen unproductive user behavior', () => {
-  test('go to PasskeyBenefits', async ({ signupFlow, page }) => {
-    await signupFlow.initializeCDPSession();
-    await signupFlow.addWebAuthn(true);
-    await signupFlow.loadAuth();
-
-    await signupFlow.navigateToPasskeyAppendScreen();
-
-    await page.getByText('Passkeys').click();
-    await signupFlow.checkLandedOnScreen(ScreenNames.PasskeyBenefits);
-  });
-
-  test('cancelling passkey input goes to LoggedIn', async ({ signupFlow, page }) => {
-    await signupFlow.initializeCDPSession();
-    await signupFlow.addWebAuthn(false);
-    await signupFlow.loadAuth();
-
-    await signupFlow.navigateToPasskeyAppendScreen();
-
-    await page.getByRole('button', { name: 'Activate' }).click();
-    await signupFlow.inputPasskey(async () => {
-      await signupFlow.checkLandedOnScreen(ScreenNames.End);
-      await signupFlow.checkNoPasskeyRegistered();
-    });
-  });
-
-  test('declining goes to LoggedIn', async ({ signupFlow, page }) => {
-    await signupFlow.initializeCDPSession();
-    await signupFlow.addWebAuthn(true);
-    await signupFlow.loadAuth();
-
-    await signupFlow.navigateToPasskeyAppendScreen();
-
-    await page.getByRole('button', { name: 'Maybe later' }).click();
-    await signupFlow.checkLandedOnScreen(ScreenNames.End);
-  });
-});
