+++ packages/tests-e2e/src/ui/default/all-browsers/signup/EmailOTP-functionality.spec.ts
@@ -1,0 +1,13 @@
+import { test } from '../../../../fixtures/UISignupTest';
+import { ScreenNames } from '../../../../utils/constants';
+
+test.describe('Signup with email OTP proper user behavior', () => {
+  test('with no passkey support', async ({ signupFlow }) => {
+    await signupFlow.createAccount();
+    await signupFlow.checkLandedOnScreen(ScreenNames.EnterOtp);
+
+    await signupFlow.fillOTP();
+    await signupFlow.checkLandedOnScreen(ScreenNames.End);
+    await signupFlow.checkNoPasskeyRegistered();
+  });
+});
