+++ packages/tests-e2e/src/ui/noName/setup.ts
@@ -1,0 +1,7 @@
+import { test as setup } from '@playwright/test';
+
+import { setBackendConfigs } from '../../utils/helperFunctions/setBackendConfigs';
+
+setup('set default configs', async () => {
+  await setBackendConfigs({ userFullNameRequired: false });
+});
