+++ packages/tests-e2e/src/ui/verificationAtSignup/chromium/signup/EmailOTP-functionality.spec.ts
@@ -1,0 +1,20 @@
+import { test } from '../../../../fixtures/UISignupTest';
+import { ScreenNames } from '../../../../utils/constants';
+
+test.describe('Signup with email OTP proper user behavior', () => {
+  test('with passkey support', async ({ signupFlow, page }) => {
+    await signupFlow.initializeCDPSession();
+    await signupFlow.addWebAuthn(true);
+    await signupFlow.loadAuth();
+
+    await signupFlow.createAccount();
+    await signupFlow.checkLandedOnScreen(ScreenNames.EnterOtp);
+
+    await signupFlow.fillOTP();
+    await signupFlow.checkLandedOnScreen(ScreenNames.PasskeyAppend);
+
+    await page.getByRole('button', { name: 'Maybe later' }).click();
+    await signupFlow.checkLandedOnScreen(ScreenNames.End);
+    await signupFlow.checkNoPasskeyRegistered();
+  });
+});
